+++ R/accessor-methods.R
@@ -14,6 +14,17 @@ setMethod('classification', signature = 'ClassyFire',
             object@classification
           })
 
+#' classification
+#' @rdname classification
+#' @description Get the `ClassyFire` classification results
+#' @param object a `Query` S4 object
+#' @export
+
+setMethod('classification', signature = 'Query',
+          function(object) {
+            object@classification
+          })
+
 #' meta
 #' @rdname meta
 #' @description Get the `ClassyFire` meta data
@@ -24,6 +35,17 @@ setMethod('meta', signature = 'ClassyFire',
           function(object) {
             object@meta
           })
+
+#' meta
+#' @rdname meta
+#' @description Get the `ClassyFire` meta data
+#' @param object a `Query` S4 object
+#' @export
+setMethod('meta', signature = 'Query',
+          function(object) {
+            object@meta
+          })
+
 
 #' description
 #' @rdname description
@@ -36,6 +58,18 @@ setMethod('description', signature = 'ClassyFire',
             object@description
           })
 
+#' description
+#' @rdname description
+#' @description Get the `ClassyFire` description
+#' @param object a `Query` S4 object
+#' @export
+#'
+setMethod('description', signature = 'Query',
+          function(object) {
+            object@description
+          })
+
+
 #' chebi
 #' @rdname chebi
 #' @description Get the predicted ChEBI identifications
@@ -43,6 +77,17 @@ setMethod('description', signature = 'ClassyFire',
 #' @export
 
 setMethod('chebi', signature = 'ClassyFire',
+          function(object) {
+            object@predicted_chebi
+          })
+
+#' chebi
+#' @rdname chebi
+#' @description Get the predicted ChEBI identifications
+#' @param object a `Query` S4 object
+#' @export
+#'
+setMethod('chebi', signature = 'Query',
           function(object) {
             object@predicted_chebi
           })
@@ -58,6 +103,17 @@ setMethod('descriptors', signature = 'ClassyFire',
             object@external_descriptors
           })
 
+#' descriptors
+#' @rdname descriptors
+#' @description Get the external descriptors
+#' @param object a `Query` S4 object
+#' @export
+
+setMethod('descriptors', signature = 'Query',
+          function(object) {
+            object@external_descriptors
+          })
+
 #' alternative_parents
 #' @rdname alternative_parents
 #' @description Get the alternative_parents
@@ -69,6 +125,18 @@ setMethod('alternative_parents', signature = 'ClassyFire',
             object@alternative_parents
           })
 
+#' alternative_parents
+#' @rdname alternative_parents
+#' @description Get the alternative_parents
+#' @param object a `Query` S4 object
+#' @export
+
+setMethod('alternative_parents', signature = 'Query',
+          function(object) {
+            object@alternative_parents
+          })
+
+
 #' direct_parent
 #' @rdname direct_parent
 #' @description Get the direct_parent
@@ -76,6 +144,17 @@ setMethod('alternative_parents', signature = 'ClassyFire',
 #' @export
 
 setMethod('direct_parent', signature = 'ClassyFire',
+          function(object) {
+            object@direct_parent
+          })
+
+#' direct_parent
+#' @rdname direct_parent
+#' @description Get the direct_parent
+#' @param object a `Query` S4 object
+#' @export
+
+setMethod('direct_parent', signature = 'Query',
           function(object) {
             object@direct_parent
           })